Fix: L2 Switch interrupt handler is called twice for each packet
When the l2 switch interrupt raises, the frame is read but
the interrupt is not cleared. The interrupt is cleared when the
handler is called the second time. This patch removes the need for
a second interrupt handler call by enabling the interrupt
only if there are no frames pending and the previous interrupt was
cleared.

Signed-off-by: Codrin Ciubotariu <codrin.ciubotariu@freescale.com>
Change-Id: I05c2b363de3e559a504efc99f224af6cd1ad3693
Reviewed-on: http://git.am.freescale.net:8181/26477
Reviewed-by: Razvan Stefanescu <razvan.stefanescu@freescale.com>
Tested-by: Razvan Stefanescu <razvan.stefanescu@freescale.com>