Add support for interrupts on PHYs connected to seville's ports
This patch creates a character device for each PHY that has
interrupt available. On probing, if the PHY nodes have interrupts
declared, each interrupt alone is registered (as shared) and
a character device is created per each PHY. The character device
can be used by user-space application to detect when a PHY interrupt
occurs. Calling open() and read()/select() will block the user-space
thread until the PHY interrupt is risen. read() will return the number
of IRQs detected since the opening of the character device. Closing
the character device will reset this number to 0.

Signed-off-by: Codrin Ciubotariu <codrin.ciubotariu@freescale.com>
Change-Id: Ibe15f46bb99042009f5479a8c77592fd928e4b98
Reviewed-on: http://git.am.freescale.net:8181/30775
Reviewed-by: Razvan Stefanescu <razvan.stefanescu@freescale.com>
Tested-by: Razvan Stefanescu <razvan.stefanescu@freescale.com>